+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/crypto/algorithms/IMXEncrypting.kt
@@ -62,4 +62,11 @@ internal interface IMXEncrypting {
                            userId: String,
                            deviceId: String,
                            senderKey: String): Boolean
+
+    /**
+     * Ensure the outbound session
+     *
+     * @param usersInRoom the users in the room
+     */
+    suspend fun ensureOutboundSession(usersInRoom: List<String>)
 }